Win64 build. Use http://annwvyn.org/dl/http://annwvyn.org/dl/Dependencies_AnnwvynSDK_64-OVR0.6.0.0-beta.zip SDK complete SDK here : http://annwvyn.org/dl/setupAnnwvynSDKx64.exe

+++ src/AnnEngine.cpp
@@ -333,7 +333,7 @@ bool AnnEngine::destroyGameObject(Annwvyn::AnnGameObject* object)
 			size_t nbObject(node->numAttachedObjects());
 			std::vector<Ogre::MovableObject*> attachedObject;
 
-			for(size_t i(0); i < nbObject; i++)
+			for(unsigned short i(0); i < nbObject; i++)
 				attachedObject.push_back(node->getAttachedObject(i));
 
 			node->detachAllObjects();
